tests: get rid of obsolete rest gw code and tests
closes #1170
