+++ pytest_tests/tests/services/rest_gate/test_rest_gate.py
@@ -4,18 +4,17 @@ import os
 import time
 
 import allure
+import neofs_env.neofs_epoch as neofs_epoch
 import pytest
-from helpers.container import create_container
+from helpers.container import create_container, delete_container, list_containers, wait_for_container_deletion
 from helpers.file_helper import generate_file, generate_file_with_content
 from helpers.neofs_verbs import put_object_to_random_node
 from helpers.rest_gate import (
     attr_into_header,
+    get_epoch_duration_via_rest_gate,
     get_object_by_attr_and_verify_hashes,
     get_via_rest_gate,
     head_object_by_attr_and_verify,
-    new_attr_into_header,
-    new_upload_via_rest_gate,
-    quote,
     try_to_get_object_and_expect_error,
     upload_via_rest_gate,
 )
@@ -25,7 +24,7 @@ from rest_gw.rest_base import TestNeofsRestBase
 from rest_gw.rest_utils import get_object_and_verify_hashes
 
 logger = logging.getLogger("NeoLogger")
-OBJECT_NOT_FOUND_ERROR = "not found"
+OBJECT_NOT_FOUND_ERROR = "404"
 
 
 @allure.link(
@@ -49,17 +48,22 @@ class TestRestGate(TestNeofsRestBase):
             "wallet_path": self.neofs_env.rest_gw.wallet.path,
         }
 
-    @allure.title("Test Put over gRPC, Get over HTTP")
+    @pytest.fixture
+    def fresh_epoch(self):
+        self.ensure_fresh_epoch()
+        yield
+
+    @allure.title("Test Put over gRPC, Get over REST")
     @pytest.mark.complex
     @pytest.mark.simple
-    def test_put_grpc_get_http(self, gw_endpoint):
+    def test_put_grpc_get_rest(self, gw_endpoint):
         """
-        Test that object can be put using gRPC interface and get using HTTP.
+        Test that object can be put using gRPC interface and get using REST gateway.
 
         Steps:
         1. Create simple and large objects.
         2. Put objects using gRPC (neofs-cli).
-        3. Download objects using HTTP gate (https://github.com/nspcc-dev/neofs-http-gw#downloading).
+        3. Download objects using REST gate (https://github.com/nspcc-dev/neofs-rest-gw).
         4. Get objects using gRPC (neofs-cli).
         5. Compare hashes for got objects.
         6. Compare hashes for got and original objects.
@@ -129,12 +133,10 @@ class TestRestGate(TestNeofsRestBase):
                 cid=cid,
                 oid=oid,
                 endpoint=gw_params["endpoint"],
-                request_path=f"/get/{cid}/{oid}",
                 return_response=True,
             )
-            content_disposition_type, filename = resp.headers["Content-Disposition"].split(";")
-            assert content_disposition_type.strip() == "inline"
-            assert filename.strip().split("=")[1] == file_path.split("/")[-1]
+
+            assert "inline" in resp.headers["Content-Disposition"]
 
         with allure.step("Verify Content-Disposition with download=true"):
             file_path = generate_file(self.neofs_env.get_object_size("simple_object_size"))
@@ -148,13 +150,11 @@ class TestRestGate(TestNeofsRestBase):
                 cid=cid,
                 oid=oid,
                 endpoint=gw_params["endpoint"],
-                request_path=f"/get/{cid}/{oid}",
                 return_response=True,
                 download=True,
             )
-            content_disposition_type, filename = resp.headers["Content-Disposition"].split(";")
-            assert content_disposition_type.strip() == "attachment"
-            assert filename.strip().split("=")[1] == file_path.split("/")[-1]
+
+            assert "attachment" in resp.headers["Content-Disposition"]
 
     @allure.title("Verify Content-Type if uploaded without any Content-Type specified")
     @pytest.mark.simple
@@ -180,7 +180,6 @@ class TestRestGate(TestNeofsRestBase):
                 cid=cid,
                 oid=oid,
                 endpoint=gw_params["endpoint"],
-                request_path=f"/get/{cid}/{oid}",
                 return_response=True,
             )
             assert resp.headers["Content-Type"] == "application/octet-stream"
@@ -198,7 +197,6 @@ class TestRestGate(TestNeofsRestBase):
                 cid=cid,
                 oid=oid,
                 endpoint=gw_params["endpoint"],
-                request_path=f"/get/{cid}/{oid}",
                 return_response=True,
             )
             assert resp.headers["Content-Type"] == "text/plain; charset=utf-8"
@@ -217,7 +215,7 @@ class TestRestGate(TestNeofsRestBase):
         with allure.step("Upload object with X-Attribute-Content-Type"):
             file_path = generate_file(self.neofs_env.get_object_size("simple_object_size"))
 
-            headers = {"X-Attribute-Content-Type": "CoolContentType"}
+            headers = {"X-Attributes": '{"Content-Type":"CoolContentType"}'}
             oid = upload_via_rest_gate(
                 cid=cid,
                 path=file_path,
@@ -229,7 +227,6 @@ class TestRestGate(TestNeofsRestBase):
                 cid=cid,
                 oid=oid,
                 endpoint=gw_params["endpoint"],
-                request_path=f"/get/{cid}/{oid}",
                 return_response=True,
             )
             assert resp.headers["Content-Type"] == "CoolContentType"
@@ -258,7 +255,6 @@ class TestRestGate(TestNeofsRestBase):
                 cid=cid,
                 oid=oid,
                 endpoint=gw_params["endpoint"],
-                request_path=f"/get/{cid}/{oid}",
                 return_response=True,
             )
             assert resp.headers["Content-Type"] == "application/json"
@@ -285,7 +281,6 @@ class TestRestGate(TestNeofsRestBase):
             cid=cid,
             oid=oid,
             endpoint=gw_params["endpoint"],
-            request_path=f"/get/{cid}/{oid}",
             return_response=True,
         )
         with open(gw_params["wallet_path"]) as wallet_file:
@@ -295,19 +290,18 @@ class TestRestGate(TestNeofsRestBase):
         assert resp.headers["X-Object-Id"] == oid
         assert resp.headers["X-Container-Id"] == cid
 
-    @allure.link("https://github.com/nspcc-dev/neofs-http-gw#uploading", name="uploading")
-    @allure.link("https://github.com/nspcc-dev/neofs-http-gw#downloading", name="downloading")
-    @allure.title("Test Put over HTTP, Get over HTTP")
+    @allure.link("https://github.com/nspcc-dev/neofs-rest-gw#docs", name="docs")
+    @allure.title("Test Put over REST, Get over REST")
     @pytest.mark.complex
     @pytest.mark.simple
-    def test_put_http_get_http(self, gw_params):
+    def test_put_rest_get_rest(self, gw_params):
         """
-        Test that object can be put and get using HTTP interface.
+        Test that object can be put and get using REST interface.
 
         Steps:
         1. Create simple and large objects.
-        2. Upload objects using HTTP (https://github.com/nspcc-dev/neofs-http-gw#uploading).
-        3. Download objects using HTTP gate (https://github.com/nspcc-dev/neofs-http-gw#downloading).
+        2. Upload objects using REST (https://github.com/nspcc-dev/neofs-rest-gw#docs).
+        3. Download objects using REST gate (https://github.com/nspcc-dev/neofs-rest-gw#docs).
         4. Compare hashes for got and original objects.
 
         Expected result:
@@ -325,7 +319,7 @@ class TestRestGate(TestNeofsRestBase):
             generate_file(self.neofs_env.get_object_size("complex_object_size")),
         )
 
-        with allure.step("Put objects using HTTP"):
+        with allure.step("Put objects using REST"):
             oid_simple = upload_via_rest_gate(cid=cid, path=file_path_simple, endpoint=gw_params["endpoint"])
             oid_large = upload_via_rest_gate(cid=cid, path=file_path_large, endpoint=gw_params["endpoint"])
 
@@ -340,25 +334,48 @@ class TestRestGate(TestNeofsRestBase):
                 endpoint=gw_params["endpoint"],
             )
 
-    @allure.link("https://github.com/nspcc-dev/neofs-http-gw#by-attributes", name="download by attributes")
-    @allure.title("Test Put over HTTP, Get over HTTP with headers")
+    @allure.link("https://github.com/nspcc-dev/neofs-rest-gw#docs", name="docs")
+    @allure.title("Test Put over REST, Get over REST with headers")
     @pytest.mark.parametrize(
         "attributes",
         [
             {"File-Name": "simple obj filename"},
             {"FileName": "simple obj filename"},
+            {"Filename": "simple_obj_filename"},
+            {"Filename": "simple_obj_filename\n"},
+            {"Filename\n": "simple_obj_filename"},
+            {"\n": "\n"},
+            {
+                """
+            key
+            """: """
+            value
+            """
+            },
+            {"\x01" * 8: "simple_obj_filename"},
+            {"Some key": "\x01" * 8},
         ],
-        ids=["hyphen", "simple"],
+        ids=[
+            "simple",
+            "hyphen",
+            "special",
+            "linebreak in value",
+            "linebreak in key",
+            "linebreaks in key and value",
+            "other linebreaks in key and value",
+            "not-so-zero bytes key",
+            "not-so-zero bytes value",
+        ],
     )
     @pytest.mark.simple
-    def test_put_http_get_http_with_headers(self, attributes: dict, gw_params):
+    def test_put_rest_get_rest_with_headers(self, attributes: dict, gw_params):
         """
         Test that object can be downloaded using different attributes in HTTP header.
 
         Steps:
         1. Create simple and large objects.
-        2. Upload objects using HTTP with particular attributes in the header.
-        3. Download objects by attributes using HTTP gate (https://github.com/nspcc-dev/neofs-http-gw#by-attributes).
+        2. Upload objects using REST gate with particular attributes in the header.
+        3. Download objects by attributes using REST gate (https://github.com/nspcc-dev/neofs-rest-gw#docs).
         4. Compare hashes for got and original objects.
 
         Expected result:
@@ -373,7 +390,7 @@ class TestRestGate(TestNeofsRestBase):
         )
         file_path = generate_file(self.neofs_env.get_object_size("simple_object_size"))
 
-        with allure.step("Put objects using HTTP with attribute"):
+        with allure.step("Put objects using REST with attribute"):
             headers = attr_into_header(attributes)
             oid = upload_via_rest_gate(
                 cid=cid,
@@ -382,16 +399,184 @@ class TestRestGate(TestNeofsRestBase):
                 endpoint=gw_params["endpoint"],
             )
 
-        attr_name = list(attributes.keys())[0]
-        attr_value = quote(str(attributes.get(attr_name)))
         get_object_by_attr_and_verify_hashes(
             oid=oid,
             file_name=file_path,
             cid=cid,
             attrs=attributes,
             endpoint=gw_params["endpoint"],
-            request_path=f"/get/{cid}/{oid}",
-            request_path_attr=f"/get_by_attribute/{cid}/{quote(str(attr_name))}/{attr_value}",
+        )
+
+    @allure.title("Test Expiration-Epoch in HTTP header")
+    @pytest.mark.simple
+    def test_expiration_epoch_in_rest(self, gw_params, fresh_epoch):
+        endpoint = self.neofs_env.sn_rpc
+
+        cid = create_container(
+            self.wallet.path,
+            shell=self.shell,
+            endpoint=endpoint,
+            rule=self.PLACEMENT_RULE_2,
+            basic_acl=PUBLIC_ACL,
+        )
+        file_path = generate_file(self.neofs_env.get_object_size("simple_object_size"))
+        oids = []
+
+        curr_epoch = neofs_epoch.get_epoch(self.neofs_env)
+        epochs = (curr_epoch, curr_epoch + 1, curr_epoch + 2, curr_epoch + 100)
+
+        for epoch in epochs:
+            headers = {"X-Attributes": '{"__NEOFS__EXPIRATION_EPOCH": "' + str(epoch) + '"}'}
+
+            with allure.step("Put objects using REST with attribute Expiration-Epoch"):
+                oids.append(
+                    upload_via_rest_gate(cid=cid, path=file_path, headers=headers, endpoint=gw_params["endpoint"])
+                )
+
+        assert len(oids) == len(epochs), "Expected all objects have been put successfully"
+
+        with allure.step("All objects can be get"):
+            for oid in oids:
+                get_via_rest_gate(cid=cid, oid=oid, endpoint=gw_params["endpoint"])
+
+        for expired_objects, not_expired_objects in [(oids[:1], oids[1:]), (oids[:2], oids[2:])]:
+            self.tick_epochs_and_wait(1)
+
+            # Wait for GC, because object with expiration is counted as alive until GC removes it
+            wait_for_gc_pass_on_storage_nodes()
+
+            for oid in expired_objects:
+                try_to_get_object_and_expect_error(
+                    cid=cid,
+                    oid=oid,
+                    error_pattern=OBJECT_NOT_FOUND_ERROR,
+                    endpoint=gw_params["endpoint"],
+                )
+
+            with allure.step("Other objects can be get"):
+                for oid in not_expired_objects:
+                    get_via_rest_gate(cid=cid, oid=oid, endpoint=gw_params["endpoint"])
+
+    @allure.title("Test other Expiration-Epoch settings in HTTP header")
+    @pytest.mark.simple
+    def test_expiration_headers_in_rest(self, gw_params, fresh_epoch):
+        endpoint = self.neofs_env.sn_rpc
+
+        cid = create_container(
+            self.wallet.path,
+            shell=self.shell,
+            endpoint=endpoint,
+            rule=self.PLACEMENT_RULE_2,
+            basic_acl=PUBLIC_ACL,
+        )
+        file_path = generate_file(self.neofs_env.get_object_size("simple_object_size"))
+        oids = []
+
+        epoch_duration = get_epoch_duration_via_rest_gate(endpoint=gw_params["endpoint"])
+        current_ts = int(time.time())
+        expiration_ts = current_ts + epoch_duration
+        expiration_rfc3339 = time.strftime("%Y-%m-%dT%H:%M:%SZ", time.gmtime(expiration_ts))
+
+        headers = (
+            {"X-Neofs-Expiration-Timestamp": str(expiration_ts)},
+            {"X-Neofs-Expiration-Duration": str(epoch_duration) + "s"},
+            {"X-Neofs-Expiration-RFC3339": expiration_rfc3339},
+        )
+
+        for header in headers:
+            keys_list = ",".join(header.keys())
+            with allure.step("Put objects using REST with attributes: " + keys_list):
+                oids.append(
+                    upload_via_rest_gate(cid=cid, path=file_path, headers=header, endpoint=gw_params["endpoint"])
+                )
+
+        assert len(oids) == len(headers), "Expected all objects have been put successfully"
+
+        with allure.step("All objects can be get"):
+            for oid in oids:
+                get_via_rest_gate(cid=cid, oid=oid, endpoint=gw_params["endpoint"])
+
+        # Wait 2 epochs because REST gate rounds time to the next epoch.
+        self.tick_epochs_and_wait(2)
+        wait_for_gc_pass_on_storage_nodes()
+
+        for oid in oids:
+            try_to_get_object_and_expect_error(
+                cid=cid,
+                oid=oid,
+                error_pattern=OBJECT_NOT_FOUND_ERROR,
+                endpoint=gw_params["endpoint"],
+            )
+
+    @allure.title("Test Negative Expiration-Epoch settings in HTTP header")
+    @pytest.mark.simple
+    def test_negative_expiration_headers_in_rest(self, gw_params, fresh_epoch):
+        endpoint = self.neofs_env.sn_rpc
+
+        cid = create_container(
+            self.wallet.path,
+            shell=self.shell,
+            endpoint=endpoint,
+            rule=self.PLACEMENT_RULE_2,
+            basic_acl=PUBLIC_ACL,
+        )
+        file_path = generate_file(self.neofs_env.get_object_size("simple_object_size"))
+
+        headers = (
+            {"X-Neofs-Expiration-Timestamp": "1635075727"},
+            {"X-Neofs-Expiration-Duration": "-1h"},
+            {"X-Neofs-Expiration-RFC3339": "2021-11-22T09:55:49Z"},
+            {"X-Neofs-Expiration-Epoch": str(neofs_epoch.get_epoch(self.neofs_env) - 1)},
+        )
+
+        for header in headers:
+            keys_list = ",".join(header.keys())
+            error_pattern = ""
+            if "Expiration-Epoch" in keys_list:
+                error_pattern = "object has expired"
+            elif "Expiration-Duration" in keys_list:
+                error_pattern = "must be positive"
+            else:
+                error_pattern = "must be in the future"
+
+            with allure.step("Put objects using REST with attributes: " + keys_list):
+                upload_via_rest_gate(
+                    cid=cid,
+                    path=file_path,
+                    headers=header,
+                    endpoint=gw_params["endpoint"],
+                    error_pattern=error_pattern,
+                )
+
+    @allure.title("Test Put over REST, Get over REST for large object")
+    @pytest.mark.complex
+    def test_put_rest_get_rest_large_file(self, gw_params):
+        """
+        This test checks upload and download with 'large' object.
+        Large is object with size up to 20Mb.
+        """
+        cid = create_container(
+            self.wallet.path,
+            shell=self.shell,
+            endpoint=self.neofs_env.sn_rpc,
+            rule=self.PLACEMENT_RULE_2,
+            basic_acl=PUBLIC_ACL,
+        )
+
+        obj_size = int(os.getenv("BIG_OBJ_SIZE", self.neofs_env.get_object_size("complex_object_size")))
+        file_path = generate_file(obj_size)
+
+        with allure.step("Put objects using REST"):
+            oid_gate = upload_via_rest_gate(cid=cid, path=file_path, endpoint=gw_params["endpoint"])
+
+        get_object_and_verify_hashes(
+            oid=oid_gate,
+            file_name=file_path,
+            wallet=self.wallet.path,
+            cid=cid,
+            shell=self.shell,
+            nodes=self.neofs_env.storage_nodes,
+            endpoint=gw_params["endpoint"],
         )
 
     @allure.title("Object with latest timestamp is returned when get by same attribute")
@@ -420,9 +605,9 @@ class TestRestGate(TestNeofsRestBase):
                 attrs = {attr_name: attr_value}
                 if not default_timestamp:
                     attrs["Timestamp"] = str(int(time.time()))
-                headers = new_attr_into_header(attrs)
+                headers = attr_into_header(attrs)
                 oids.append(
-                    new_upload_via_rest_gate(cid=cid, path=file_path, endpoint=gw_params["endpoint"], headers=headers)
+                    upload_via_rest_gate(cid=cid, path=file_path, endpoint=gw_params["endpoint"], headers=headers)
                 )
                 time.sleep(1)
 
@@ -442,9 +627,9 @@ class TestRestGate(TestNeofsRestBase):
                 attrs = {attr_name: attr_value, attr_name2: attr_value2}
                 if not default_timestamp:
                     attrs["Timestamp"] = str(int(time.time()))
-                headers = new_attr_into_header(attrs)
+                headers = attr_into_header(attrs)
                 oids.append(
-                    new_upload_via_rest_gate(
+                    upload_via_rest_gate(
                         cid=cid,
                         path=file_path,
                         headers=headers,
@@ -488,9 +673,9 @@ class TestRestGate(TestNeofsRestBase):
                 attrs = {attr_name: attr_value}
                 if not default_timestamp:
                     attrs["Timestamp"] = str(int(time.time()))
-                headers = new_attr_into_header(attrs)
+                headers = attr_into_header(attrs)
                 oids.append(
-                    new_upload_via_rest_gate(
+                    upload_via_rest_gate(
                         cid=cid,
                         path=file_path,
                         headers=headers,
@@ -514,9 +699,9 @@ class TestRestGate(TestNeofsRestBase):
                 attrs = {attr_name: attr_value, attr_name2: attr_value2}
                 if not default_timestamp:
                     attrs["Timestamp"] = str(int(time.time()))
-                headers = new_attr_into_header(attrs)
+                headers = attr_into_header(attrs)
                 oids.append(
-                    new_upload_via_rest_gate(
+                    upload_via_rest_gate(
                         cid=cid,
                         path=file_path,
                         headers=headers,
@@ -533,90 +718,133 @@ class TestRestGate(TestNeofsRestBase):
                 endpoint=gw_params["endpoint"],
             )
 
-    @allure.title("Test Expiration-Epoch in HTTP header")
-    @pytest.mark.simple
-    def test_expiration_epoch_in_http(self, gw_params):
-        endpoint = self.neofs_env.sn_rpc
+    @allure.title("Test Put over gRPC, Get over HTTP")
+    @pytest.mark.parametrize(
+        "object_size",
+        [
+            pytest.param("simple_object_size", id="simple object", marks=pytest.mark.simple),
+            pytest.param("complex_object_size", id="complex object", marks=pytest.mark.complex),
+        ],
+    )
+    def test_object_put_get_attributes(self, object_size: int, gw_params: dict):
+        """
+        Test that object can be put using gRPC interface and get using HTTP.
 
+        Steps:
+        1. Create object;
+        2. Put objects using gRPC (neofs-cli) with attributes [--attributes chapter1=peace,chapter2=war];
+        3. Download object;
+        4. Compare hashes between original and downloaded object;
+        5. Download the object with specified attributes;
+        6. Compare hashes between original and downloaded object;
+
+
+        Expected result:
+        Hashes must be the same.
+        """
+        gw_attributes = {
+            "obj_key1": "Chapter1",
+            "obj_value1": "peace",
+            "obj_key2": "Chapter2",
+            "obj_value2": "war",
+        }
+
+        with allure.step("Create public container"):
+            cid = create_container(
+                self.wallet.path,
+                shell=self.shell,
+                endpoint=self.neofs_env.sn_rpc,
+                rule=self.PLACEMENT_RULE_1,
+                basic_acl=PUBLIC_ACL,
+            )
+
+        file_path = generate_file(self.neofs_env.get_object_size(object_size))
+
+        obj_key1 = gw_attributes["obj_key1"]
+        obj_value1 = gw_attributes["obj_value1"]
+        obj_key2 = gw_attributes["obj_key2"]
+        obj_value2 = gw_attributes["obj_value2"]
+
+        key_value1 = obj_key1 + "=" + obj_value1
+        key_value2 = obj_key2 + "=" + obj_value2
+
+        with allure.step("Put objects using gRPC [--attributes chapter1=peace,chapter2=war]"):
+            oid = put_object_to_random_node(
+                wallet=self.wallet.path,
+                path=file_path,
+                cid=cid,
+                shell=self.shell,
+                neofs_env=self.neofs_env,
+                attributes=f"{key_value1},{key_value2}",
+            )
+        with allure.step("Get object and verify hashes"):
+            get_object_and_verify_hashes(
+                oid=oid,
+                file_name=file_path,
+                wallet=self.wallet.path,
+                cid=cid,
+                shell=self.shell,
+                nodes=self.neofs_env.storage_nodes,
+                endpoint=gw_params["endpoint"],
+            )
+
+        attrs = {obj_key1: obj_value1, obj_key2: obj_value2}
+
+        with allure.step("Download the object with attribute [objects/$CID/by_attribute/chapter1/peace]"):
+            get_object_by_attr_and_verify_hashes(
+                oid=oid,
+                file_name=file_path,
+                cid=cid,
+                attrs=attrs,
+                endpoint=gw_params["endpoint"],
+            )
+
+        with allure.step(
+            "[Negative] Try to get object with invalid attribute [objects/$CID/by_attribute/chapter1/war]"
+        ):
+            with pytest.raises(Exception, match=".*object not found.*"):
+                get_object_by_attr_and_verify_hashes(
+                    oid=oid,
+                    file_name=file_path,
+                    cid=cid,
+                    attrs={obj_key1: obj_value2},
+                    endpoint=gw_params["endpoint"],
+                )
+
+    def test_get_object_after_container_deleted(self, gw_params: dict):
         cid = create_container(
-            self.wallet.path,
+            wallet=self.wallet.path,
             shell=self.shell,
-            endpoint=endpoint,
-            rule=self.PLACEMENT_RULE_2,
+            endpoint=self.neofs_env.sn_rpc,
+            rule=self.PLACEMENT_RULE_1,
             basic_acl=PUBLIC_ACL,
         )
         file_path = generate_file(self.neofs_env.get_object_size("simple_object_size"))
-        oids = []
-
-        curr_epoch = self.ensure_fresh_epoch()
-        epochs = (curr_epoch, curr_epoch + 1, curr_epoch + 2, curr_epoch + 100)
-
-        for epoch in epochs:
-            headers = {"X-Attribute-Neofs-Expiration-Epoch": str(epoch)}
-
-            with allure.step("Put objects using HTTP with attribute Expiration-Epoch"):
-                oids.append(
-                    upload_via_rest_gate(cid=cid, path=file_path, headers=headers, endpoint=gw_params["endpoint"])
-                )
-
-        assert len(oids) == len(epochs), "Expected all objects have been put successfully"
-
-        with allure.step("All objects can be get"):
-            for oid in oids:
-                get_via_rest_gate(
-                    cid=cid,
-                    oid=oid,
-                    endpoint=gw_params["endpoint"],
-                    request_path=f"/get/{cid}/{oid}",
-                )
-
-        for expired_objects, not_expired_objects in [(oids[:1], oids[1:]), (oids[:2], oids[2:])]:
-            self.tick_epochs_and_wait(1)
-
-            # Wait for GC, because object with expiration is counted as alive until GC removes it
-            wait_for_gc_pass_on_storage_nodes()
-
-            for oid in expired_objects:
-                try_to_get_object_and_expect_error(
-                    cid=cid,
-                    oid=oid,
-                    error_pattern=OBJECT_NOT_FOUND_ERROR,
-                    endpoint=gw_params["endpoint"],
-                )
-
-            with allure.step("Other objects can be get"):
-                for oid in not_expired_objects:
-                    get_via_rest_gate(
-                        cid=cid, oid=oid, endpoint=gw_params["endpoint"], request_path=f"/get/{cid}/{oid}"
-                    )
-
-    @allure.title("Test Put over HTTP, Get over HTTP for large object")
-    @pytest.mark.complex
-    def test_put_http_get_http_large_file(self, gw_params):
-        """
-        This test checks upload and download with 'large' object.
-        Large is object with size up to 20Mb.
-        """
-        cid = create_container(
-            self.wallet.path,
-            shell=self.shell,
-            endpoint=self.neofs_env.sn_rpc,
-            rule=self.PLACEMENT_RULE_2,
-            basic_acl=PUBLIC_ACL,
-        )
-
-        obj_size = int(os.getenv("BIG_OBJ_SIZE", self.neofs_env.get_object_size("complex_object_size")))
-        file_path = generate_file(obj_size)
-
-        with allure.step("Put objects using HTTP"):
-            oid_gate = upload_via_rest_gate(cid=cid, path=file_path, endpoint=gw_params["endpoint"])
-
-        get_object_and_verify_hashes(
-            oid=oid_gate,
-            file_name=file_path,
-            wallet=self.wallet.path,
+        oid = upload_via_rest_gate(
             cid=cid,
-            shell=self.shell,
-            nodes=self.neofs_env.storage_nodes,
+            path=file_path,
             endpoint=gw_params["endpoint"],
         )
+
+        with allure.step("Delete container and verify container deletion"):
+            delete_container(
+                wallet=self.wallet.path,
+                cid=cid,
+                shell=self.shell,
+                endpoint=self.neofs_env.sn_rpc,
+            )
+            self.tick_epochs_and_wait(1)
+            wait_for_container_deletion(
+                self.wallet.path,
+                cid,
+                shell=self.shell,
+                endpoint=self.neofs_env.sn_rpc,
+            )
+            assert cid not in list_containers(self.wallet.path, shell=self.shell, endpoint=self.neofs_env.sn_rpc)
+        with allure.step("Try to download object after container deletion"):
+            try_to_get_object_and_expect_error(
+                cid=cid,
+                oid=oid,
+                error_pattern=OBJECT_NOT_FOUND_ERROR,
+                endpoint=gw_params["endpoint"],
+            )
